requirements.txt
captum==0.7.0
matplotlib==3.8.2
opencv_python==4.8.1.78
Pillow==10.1.0
scikit_learn==1.3.2
spacy==3.7.2
torch==2.1.2+cu121
torchvision==0.16.2+cu121
tqdm==4.66.1
transformers==4.36.2
